lu to d and add (int)

+++ main.cpp
@@ -59,7 +59,7 @@ bool is_normal_zh_symbol(char a_char_0, char a_char_1, char a_char_2)
 
 void split_word(string sentence)
 {
-	printf("%s - %lu\n", sentence.c_str(), sentence.length());
+	printf("%s - %d\n", sentence.c_str(), (int)sentence.length());
 }
 
 int main()
@@ -101,7 +101,7 @@ int main()
 					split_word(sentence);
 				}
 				sentence = "";
-				printf("x -- %lu\n", sentence.length());
+				printf("x -- %d\n", (int)sentence.length());
 			} else if (is_normal_zh_symbol(zh_char[0], zh_char[1], zh_char[2])) {
 //				printf("zn - %c%c%c\n", zh_char[0], zh_char[1], zh_char[2]);
 			} else {
@@ -116,7 +116,7 @@ int main()
 					split_word(sentence);
 				}
 				sentence = "";
-				printf("y -- %lu\n", sentence.length());
+				printf("y -- %d\n", (int)sentence.length());
 			} else if (is_normal_en_symbol(en_char)) {
 //				printf("cn - %c\n", en_char);
 			} else {
